requirements.txt
dj-database-url==0.4.2
Django==1.10.6
django-crispy-forms==1.6.1
django-filter==1.0.1
django-leaflet==0.21.0
djangorestframework==3.5.4
djangorestframework-gis==0.11
geoip2==2.4.2
gunicorn==19.7.0
ipaddress==1.0.18
maxminddb==1.3.0
olefile==0.44
Pillow==4.0.0
psycopg2==2.7
pytz==2016.10
requests==2.13.0
whitenoise==3.3.0
